Changed ' to ", to fix m3u/m3u8 list specific to Jellyfin loading m3u lists.
For some odd reason Jellyfin doesn't like ' around data it prefers ".  Also changed channelID to channel-id.  I see this more commonly and it also uses that id in the epg/xmltv files.

This should help deal with guide data and loading icons for channels without going in an actually mapping channels in Jellyfin.